+++ package.json
@@ -2,7 +2,7 @@
   "name": "site",
   "version": "1.0.0",
   "description": "Die Lernwerkstatt Website",
-  "main": "index.js",
+  "main": "src/index.js",
   "scripts": {
     "test": "jest --config config/jest.json",
     "start": "cross-env NODE_ENV=development && nodemon index.js",
