+++ src/routes/blogs.js
@@ -1,7 +1,7 @@
 const express = require("express");
 const mongoose = require("mongoose");
 
-const Blogposts = require("../../models/blogposts");
+const Blogposts = require("../models/blogposts");
 const { dbUrl } = require("../../config/secrets");
 
 mongoose.connect(
