+++ src/views/blogs.hbs
