+++ src/views/courses.hbs
