+++ src/views/partials/blogs/single.hbs
