+++ src/views/partials/main/footer.test.js
