+++ src/views/partials/workshops/seminar.hbs
