+++ src/views/partials/workshops/seminar.test.js
