+++ src/views/workshops.hbs
